Feature/party - S3 이미지 전체 적용

+++ src/features/banner/banner.service.ts
@@ -1,26 +1,37 @@
 import { Injectable } from '@nestjs/common';
 import { BannerRepository } from './repository/banner.repository';
-import { ImageService } from 'src/libs/image/image.service';
+import { S3Service } from 'src/libs/aws/s3/s3.service';
+import { bannerImageKey } from 'src/libs/aws/s3/key.util';
 
 @Injectable()
 export class BannerService {
   constructor(
     private bannerRepository: BannerRepository,
-    private imageService: ImageService,
+    private s3Service: S3Service,
   ) {}
 
-  async createWeb(title: string, imagePath: string, link: string) {
-    const savedImagePath = imagePath ? this.imageService.getRelativePath(imagePath) : undefined;
+  async createWeb(title: string, link: string, image: Express.Multer.File) {
+    let key: string | undefined;
 
-    const result = await this.bannerRepository.createWeb(title, savedImagePath, link);
+    if (image) {
+      key = bannerImageKey(image.originalname);
+      await this.s3Service.uploadFile(image, key);
+    }
+
+    const result = await this.bannerRepository.createWeb(title, link, image.originalname);
 
     return result;
   }
 
-  async createApp(title: string, imagePath: string, link: string) {
-    const savedImagePath = imagePath ? this.imageService.getRelativePath(imagePath) : undefined;
+  async createApp(title: string, link: string, image: Express.Multer.File) {
+    let key: string | undefined;
 
-    const result = await this.bannerRepository.createApp(title, savedImagePath, link);
+    if (image) {
+      key = bannerImageKey(image.originalname);
+      await this.s3Service.uploadFile(image, key);
+    }
+
+    const result = await this.bannerRepository.createApp(title, link, image.originalname);
 
     return result;
   }
@@ -38,6 +49,14 @@ export class BannerService {
   }
 
   async delete(id: number) {
-    await this.bannerRepository.delete(id);
+    const deleteBanner = await this.bannerRepository.findById(id);
+
+    if (deleteBanner) {
+      await this.bannerRepository.delete(id);
+
+      if (deleteBanner.image) {
+        this.s3Service.deleteFile(deleteBanner.image);
+      }
+    }
   }
 }
